Further hello-world size cuts: 4 more lazy-load splits
Builds on the previous classic-renderer/routing-keywords split. The
hello-world smoke test goes from 168.59 KB / 53.67 KB gzip to
160.11 KB / 50.88 KB gzip. Cumulative with prior commit: 243.30 KB /
77.32 KB → 160.11 KB / 50.88 KB (34.2% gzip reduction). Classic
v2-app-template stays flat (319.55 KB / 99.38 KB).

Each change is the same registration pattern as before — a separate
side-effect file imported by classic-app `setup-registry`, leaving the
heavy module out of the renderer-only path.

1. **Curly symbols extracted to `curly-symbols.ts`.** `BOUNDS`,
   `DIRTY_TAG`, `IS_DISPATCHING_ATTRS`, and the new `CURLY_COMPONENT_BRAND`
   live in their own file. `isCurlyManager` is now a brand check
   (`manager[CURLY_COMPONENT_BRAND] === true`) instead of an instance
   check, so the resolver no longer pulls in `./curly` (the full
   `CurlyComponentManager` lifecycle, ~17 KB) just to identify the
   manager. `curly.ts` re-exports the symbols for back-compat and tags
   `CURLY_COMPONENT_MANAGER` with the brand. `classic-renderer.ts` and
   `resolver.ts` now import from `curly-symbols.ts`.

2. **`@ember/-internals/glimmer/lib/component`'s top-level side effects
   moved to `register-curly-component.ts`.**
   `setInternalComponentManager(CURLY_COMPONENT_MANAGER, Component)` and
   `Component.reopenClass({ positionalParams: [] })` ran at module load
   time, which kept the full classic `Component` class graph reachable
   from anything that imported `@ember/component` (e.g.
   `@glimmer/component`'s `setComponentManager`/`capabilities` imports).
   The registration now lives in a side-effect-only file imported by
   `setup-registry.ts`, so classic apps still get it on boot.

3. **`DebugRenderTreeImpl` factory moved behind a registry in
   `@glimmer/runtime/.../environment.ts`.** Previously `EnvironmentImpl`
   imported `DebugRenderTree` statically and only constructed one when
   `delegate.enableDebugTooling` was true — but the import alone pulled
   the whole class (and its `getDebugName` cousin) into every bundle.
   New `registerDebugRenderTreeFactory` lets a side-effect module
   (`debug-render-tree-register.ts`) supply the constructor; without
   that import, `env.debugRenderTree` stays `undefined` even when the
   delegate flag is set. Classic apps re-register it via
   `setup-registry.ts`. `getDebugName` was the other static reach into
   `debug-render-tree`, so it moved to its own file (`get-debug-name.ts`)
   that opcodes can import without dragging the rest in.

4. **`to-bool.ts` swapped `isArray` from `@ember/array` for
   `Array.isArray(x) || isEmberArray(x)`.** `isArray` from
   `@ember/array` calls `EmberArray.detect`, which transitively pulls
   `@ember/array`'s entire mixin/Enumerable/Observable/computed graph
   (~16 KB) just to test array-ness inside `{{#if}}`. Using
   `isEmberArray` from `@ember/array/-internals` (a WeakSet brand set
   in `EmberArray#init`) covers all instances of EmberArray-mixed
   classes — the same set the old check covered in practice.

Verified: lint clean, all type-checks pass, `test:node` 20/20,
`test:blueprints` 265/265, both smoke-test apps build.
