+++ packages/@ember/-internals/glimmer/lib/classic-renderer.ts
@@ -35,7 +35,7 @@ import type { SimpleDocument, SimpleElement, SimpleNode } from '@simple-dom/inte
 import { hasDOM } from '../../browser-environment';
 import type Component from './component';
 import type ClassicComponent from './component';
-import { BOUNDS } from './component-managers/curly';
+import { BOUNDS } from './component-managers/curly-symbols';
 import { createRootOutlet } from './component-managers/outlet';
 import { RootComponentDefinition } from './component-managers/root';
 import { makeRouteTemplate } from './component-managers/route-template';
