+++ packages/@ember/-internals/glimmer/lib/component.ts
@@ -18,19 +18,17 @@ import { guidFor } from '@ember/-internals/utils/lib/guid';
 import { assert } from '@ember/debug';
 import { DEBUG } from '@glimmer/env';
 import type { Environment, Template, TemplateFactory } from '@glimmer/interfaces';
-import { setInternalComponentManager } from '@glimmer/manager/lib/internal/api';
 import { isUpdatableRef, updateRef } from '@glimmer/reference/lib/reference';
 import { normalizeProperty } from '@glimmer/runtime/lib/dom/props';
 import type { DirtyableTag } from '@glimmer/interfaces';
 import { createTag, DIRTY_TAG as dirtyTag } from '@glimmer/validator/lib/validators';
 import type { SimpleElement } from '@simple-dom/interface';
+import { getComponentCapturedArgs } from './component-managers/curly';
 import {
   BOUNDS,
-  CURLY_COMPONENT_MANAGER,
   DIRTY_TAG,
   IS_DISPATCHING_ATTRS,
-  getComponentCapturedArgs,
-} from './component-managers/curly';
+} from './component-managers/curly-symbols';
 import hasDOM from '@ember/-internals/browser-environment/lib/has-dom';
 
 // Keep track of which component classes have already been processed for lazy event setup.
@@ -1682,11 +1680,13 @@ class Component<S = unknown>
   }
 }
 
-// We continue to use reopenClass here so that positionalParams can be overridden with reopenClass in subclasses.
-Component.reopenClass({
-  positionalParams: [],
-});
-
-setInternalComponentManager(CURLY_COMPONENT_MANAGER, Component);
+// `Component.reopenClass({ positionalParams: [] })` and the curly component
+// manager registration that used to live here now happen in
+// `./register-curly-component`, imported as a side effect by classic-app
+// setup (`./setup-registry`). Splitting the side effect out lets bundlers
+// tree-shake the classic `Component` class out of bundles that don't
+// otherwise need it — e.g. the hello-world app that only uses
+// `setComponentManager` / `capabilities` from `@ember/component` via
+// `@glimmer/component`.
 
 export default Component;
