+++ packages/@ember/-internals/glimmer/lib/utils/to-bool.ts
@@ -1,16 +1,25 @@
 import { isHTMLSafe } from './string';
 import { get } from '@ember/-internals/metal/lib/property_get';
 import { tagForProperty } from '@ember/-internals/metal/lib/tags';
-import { isArray } from '@ember/array';
+import { isEmberArray } from '@ember/array/-internals';
 import { isProxy } from '@ember/-internals/utils/lib/is_proxy';
 import { consumeTag } from '@glimmer/validator/lib/tracking';
+
+// Lightweight array predicate — covers native arrays plus anything branded
+// via `setEmberArray` (which is what `EmberArray.init` does, so any
+// instance of an EmberArray-mixed class is branded). Avoids the full
+// `@ember/array/index` graph (Mixin, Enumerable, Observable, computed)
+// just to test array-ness for `{{#if}}`.
+function isArrayLike(obj: unknown): obj is ArrayLike<unknown> {
+  return Array.isArray(obj) || isEmberArray(obj);
+}
 
 export default function toBool(predicate: unknown): boolean {
   if (isProxy(predicate)) {
     consumeTag(tagForProperty(predicate, 'content'));
 
     return Boolean(get(predicate, 'isTruthy'));
-  } else if (isArray(predicate)) {
+  } else if (isArrayLike(predicate)) {
     consumeTag(tagForProperty(predicate as object, '[]'));
 
     return (predicate as { length: number }).length !== 0;
