+++ packages/@glimmer/runtime/lib/debug-render-tree.ts
@@ -2,7 +2,6 @@ import { DEBUG } from '@glimmer/env';
 import type {
   Bounds,
   CapturedRenderNode,
-  ComponentDefinition,
   DebugRenderTree,
   Nullable,
   RenderNode,
@@ -198,9 +197,7 @@ export default class DebugRenderTreeImpl<
   }
 }
 
-export function getDebugName(
-  definition: ComponentDefinition,
-  manager = definition.manager
-): string {
-  return definition.resolvedName ?? definition.debugName ?? manager.getDebugName(definition.state);
-}
+// `getDebugName` now lives in `./get-debug-name` so opcodes can pull it
+// in without dragging in the full `DebugRenderTreeImpl` class. Re-export
+// for back-compat with any external consumers.
+export { getDebugName } from './get-debug-name';
